New bazel macro to test run installed drake.
`install_test()` is a wrapper macro that creates a Python script that is run
during the testing phase. The Python script installs `drake` and runs the
targets that are listed in `install_tests` in install rules.
The Python script that is created will allow to test multiple targets in
the install tree without having to install drake multiple times.
`install_test()` should be called only once, in `//:*`.
All the install tests that have been created in all the `install()` commands
in the source tree will be integrated in the Python script that is generated.
The targets can either be executables that are installed, or any executable
or scripts that is designed to test drake features in the install tree.
The Python install_test_helper file (`tools/install/install_test_helper.py`)
has been improved to easily create Python scripts that run commands from the
install directory. This is convenient when trying to test an executable in
the install tree that is run with parameters, or an executable that needs to
be killed because it never stops running.